FIxed bug in onciditonal celltype analysis (up/down stream arguments were not passed along)
Added error checks to the functions which make the covar/genesets files, which catches errors when wrong specificity_species argument is used

The additional datasets were not loading. Fixed this.

Fixed a bug in load.magma.results.file (controlledCTs was not being passed as an argument, but was still being used to drop excess results)

Altered prepare.quantile.groups to use functions from EWCE